db_utils/requirements.txt
psycopg2-binary
psycopg[binary]
Faker
python-dotenv 